# backend/app/routers/auth.py
import bleach
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from sqlalchemy.orm import Session

from .. import auth, models, schemas, utils
from ..auth import ALGORITHM, SECRET_KEY
from ..database import get_db


def sanitize_str(value: str) -> str:
    return bleach.clean(value, tags=[], attributes={}, strip=True)


oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")


def get_current_user(
    token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)
):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise HTTPException(status_code=401, detail="Invalid token")
    except JWTError:
        raise HTTPException(status_code=401, detail="Invalid token")

    user = db.query(models.User).filter(models.User.username == username).first()
    if user is None:
        raise HTTPException(status_code=401, detail="User not found")
    return user


router = APIRouter(prefix="/auth", tags=["Auth"])


@router.post("/register", response_model=schemas.UserResponse, summary="Register a new user",
    description="Creates a new user account with username, email, and password. "
                "If no role is provided, the default role is **trader**."
)
def register(user: schemas.UserCreate, db: Session = Depends(get_db)):
    existing_user = (
        db.query(models.User).filter(models.User.username == user.username).first()
    )
    if existing_user:
        raise HTTPException(status_code=400, detail="Username already registered")

    hashed_pw = utils.hash_password(user.password)
    new_user = models.User(
        username=sanitize_str(user.username),
        email=sanitize_str(user.email),
        hashed_password=hashed_pw,
        role=sanitize_str(user.role) if user.role else "trader",
    )
    db.add(new_user)
    db.commit()
    db.refresh(new_user)
    return new_user


@router.post("/login", summary="User login",
    description="Authenticates a user with username and password. "
                "Returns a JWT access token if credentials are valid."
)
def login(user: schemas.UserLogin, db: Session = Depends(get_db)):
    db_user = (
        db.query(models.User).filter(models.User.username == user.username).first()
    )
    if not db_user or not utils.verify_password(user.password, db_user.hashed_password):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid credentials"
        )

    token = auth.create_access_token({"sub": db_user.username})
    return {"access_token": token, "token_type": "bearer"}




@router.get("/me", response_model=schemas.UserResponse, summary="Get current user profile",
    description="Returns details (id, username, email, role) of the currently authenticated user."
)
def read_users_me(current_user: models.User = Depends(get_current_user)):
    return current_user